# docs/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WeChat Channels Download API",
    "version": "v260613",
    "description": "可以通过 HTTP 请求来调用下载器的 API，现在提供了以下几类接口：\n\n- 视频号相关\n- 公众号相关\n- 下载器相关"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:2022",
      "description": "本地服务器"
    }
  ],
...
  "paths": {
...
    "/api/browse_history/create": {
      "post": {
        "tags": ["下载器"],
        "summary": "创建/更新浏览历史",
        "description": "根据内容 id 进行 upsert：首次写入创建记录，重复写入会将 visited_times + 1，并更新内容/账号信息。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "内容 id（视频号 object id）"
                  },
                  "contact": {
                    "type": "object",
                    "properties": {
                      "username": {
                        "type": "string",
                        "description": "视频号账号 username（如 v2_xxx@finder）"
                      },
                      "nickname": {
                        "type": "string",
                        "description": "视频号昵称"
                      },
                      "headUrl": {
                        "type": "string",
                        "description": "视频号头像 URL"
                      }
                    },
                    "required": ["username"]
                  },
                  "objectDesc": {
                    "type": "object",
                    "properties": {
                      "description": {
                        "type": "string",
                        "description": "内容描述/标题"
                      },
                      "media": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "url": {
                              "type": "string",
                              "description": "媒体直链（不含 urlToken）"
                            },
                            "coverUrl": {
                              "type": "string",
                              "description": "封面 URL"
                            },
                            "decodeKey": {
                              "type": "string",
                              "description": "解密 key（视频号下载用）"
                            },
                            "urlToken": {
                              "type": "string",
                              "description": "媒体 token（拼接到 url 后）"
                            }
                          }
                        },
                        "description": "媒体列表（取第 1 个用于写入 content_url/content_cover_url）"
                      }
                    }
                  },
                  "objectNonceId": {
                    "type": "string",
                    "description": "nonce id（写入 extra_data）"
                  },
                  "source_url": {
                    "type": "string",
                    "description": "来源页面 URL"
                  }
                },
                "required": ["id"]
              },
              "examples": {
                "create": {
                  "summary": "记录浏览历史",
                  "value": {
                    "id": "14545102784038246591",
                    "contact": {
                      "username": "v2_xxx@finder",
                      "nickname": "示例账号",
                      "headUrl": "https://wx.qlogo.cn/finderhead/ver_1/xxx/132"
                    },
                    "objectDesc": {
                      "description": "示例内容标题",
                      "media": [
                        {
                          "url": "https://finder.video.qq.com/251/20302/stodownload?encfilekey=xxx",
                          "coverUrl": "https://wx.qlogo.cn/finder/xxx/0",
                          "decodeKey": "81721112",
                          "urlToken": "&token=xxx"
                        }
                      ]
                    },
                    "objectNonceId": "nonce_id_xxx",
                    "source_url": "https://channels.weixin.qq.com/xxx"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "写入成功",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "成功",
                    "value": {
                      "code": 0,
                      "msg": "成功",
                      "data": {}
                    }
                  },
                  "missing_id": {
                    "summary": "缺少 id",
                    "value": {
                      "code": 400,
                      "msg": "缺少内容 id",
                      "data": {}
                    }
                  },
                  "invalid": {
                    "summary": "参数不合法",
                    "value": {
                      "code": 400,
                      "msg": "不合法的参数",
                      "data": {}
                    }
                  },
                  "no_db": {
                    "summary": "数据库未初始化",
                    "value": {
                      "code": 500,
                      "msg": "数据库未初始化",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/browse_history/list": {
      "post": {
        "tags": ["下载器"],
        "summary": "获取浏览历史列表",
        "description": "按 updated_at 倒序返回浏览历史。可选通过 username 过滤指定视频号账号。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "视频号账号 username（如 v2_xxx@finder）。不传则返回全部。"
                  }
                }
              },
              "examples": {
                "all": {
                  "summary": "查询全部",
                  "value": {}
                },
                "by_username": {
                  "summary": "按账号过滤",
                  "value": {
                    "username": "v2_xxx@finder"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "浏览历史列表",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "成功",
                    "value": {
                      "code": 0,
                      "msg": "成功",
                      "data": {
                        "list": [
                          {
                            "id": 1,
                            "platform_id": "wx_channels",
                            "visited_times": 3,
                            "account_id": null,
                            "influencer_id": null,
                            "account_external_id": "v2_xxx@finder",
                            "account_username": "v2_xxx@finder",
                            "account_nickname": "示例账号",
                            "account_avatar_url": "https://wx.qlogo.cn/finderhead/ver_1/xxx/132",
                            "content_id": null,
                            "content_type": "video",
                            "content_external_id": "14545102784038246591",
                            "content_title": "示例内容标题",
                            "content_url": "https://finder.video.qq.com/251/20302/stodownload?encfilekey=xxx",
                            "content_source_url": "https://channels.weixin.qq.com/xxx",
                            "content_cover_url": "https://wx.qlogo.cn/finder/xxx/0",
                            "extra_data": "{\"nonce_id\":\"nonce_id_xxx\",\"decodeKey\":\"81721112\",\"urlToken\":\"&token=xxx\"}",
                            "created_at": 1710000000000,
                            "updated_at": 1710000000000,
                            "deleted_at": null
                          }
                        ]
                      }
                    }
                  },
                  "invalid": {
                    "summary": "参数不合法",
                    "value": {
                      "code": 400,
                      "msg": "EOF",
                      "data": {}
                    }
                  },
                  "no_db": {
                    "summary": "数据库未初始化",
                    "value": {
                      "code": 500,
                      "msg": "数据库未初始化",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/show_file": {
      "post": {
        "tags": ["下载器"],
        "summary": "在文件夹中显示文件",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "path": {
                    "type": "string",
                    "description": "文件所在目录绝对路径"
                  },
                  "name": {
                    "type": "string",
                    "description": "文件名"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "成功",
                    "value": {
                      "code": 0,
                      "msg": "成功",
                      "data": {}
                    }
                  },
                  "fail": {
                    "summary": "失败",
                    "value": {
                      "code": 400,
                      "msg": "参数错误",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/download_task/list": {
      "post": {
        "tags": ["下载任务"],
        "summary": "获取下载任务列表（兼容 wx_channels_api）",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "page": { "type": "integer", "default": 1, "description": "页码（从 1 开始）" },
                  "pageSize": { "type": "integer", "default": 20, "description": "每页条数" },
                  "status": { "type": "integer", "description": "按 download_task.status 过滤" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "任务列表",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "成功",
                    "value": {
                      "code": 0,
                      "msg": "成功",
                      "data": {
                        "list": [
                          {
                            "id": 1,
                            "task_id": "fqG2OhiJFRYpnELqZmPRW",
                            "type": 1,
                            "status": 4,
                            "external_id": "14545102784038246591",
                            "protocol": "http",
                            "url": "https://finder.video.qq.com/251/20302/stodownload?encfilekey=xxx&token=xxx",
                            "title": "示例标题.mp4",
                            "cover_url": "https://wx.qlogo.cn/finder/xxx/0",
                            "size": 1462345,
                            "progress": "{\"downloaded\":1462345}",
                            "filepath": "苏苏是猪猪/示例标题.mp4",
                            "error": "",
                            "reason": "",
                            "metadata1": "{}",
                            "metadata2": "{\"platform\":\"wx_channels\"}",
                            "idx": 0,
                            "created_at": "1760000000000",
                            "updated_at": "1760000001000",
                            "events": [
                              {
                                "id": 1,
                                "task_id": 1,
                                "type": "started",
                                "message": "",
                                "data": "",
                                "created_at": "1760000000000"
                              }
                            ]
                          }
                        ],
                        "page": 1,
                        "page_size": 20,
                        "total": 1
                      }
                    }
                  },
                  "no_db": {
                    "summary": "数据库未初始化",
                    "value": { "code": 500, "msg": "数据库未初始化", "data": {} }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/download_task/start": {
      "post": {
        "tags": ["下载任务"],
        "summary": "开始/恢复下载任务（兼容 wx_channels_api）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "download_task_id": { "type": "integer", "description": "download_task 表主键 id" } },
                "required": ["download_task_id"]
              },
              "examples": { "start": { "summary": "开始任务", "value": { "download_task_id": 1 } } }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作结果",
            "content": {
              "application/json": {
                "examples": {
                  "success": { "summary": "成功", "value": { "code": 0, "msg": "成功", "data": { "message": "开始下载" } } },
                  "not_found": { "summary": "任务不存在", "value": { "code": 404, "msg": "未找到下载任务", "data": {} } }
                }
              }
            }
          }
        }
      }
    },
    "/api/download_task/profile": {
      "post": {
        "tags": ["下载任务"],
        "summary": "获取单个下载任务（兼容 wx_channels_api）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "task_id": { "type": "string", "description": "下载器 task id（download_task.task_id）" } },
                "required": ["task_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "任务详情",
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "summary": "成功",
                    "value": { "code": 0, "msg": "成功", "data": { "id": 1, "task_id": "fqG2OhiJFRYpnELqZmPRW", "status": 4, "url": "https://finder.video.qq.com/xxx", "title": "示例标题.mp4" } }
                  },
                  "not_found": { "summary": "任务不存在", "value": { "code": 404, "msg": "未找到下载任务", "data": {} } }
                }
              }
            }
          }
        }
      }
    },
    "/api/download_task/create": {
      "post": {
        "tags": ["下载任务"],
        "summary": "创建视频号下载任务（兼容 wx_channels_api）",
        "description": "接收 ChannelsObject 结构的原始视频号数据，后端自动处理 format_feed（类似 format_feed 逻辑）、链接拼接、文件名生成、spec 处理等。支持 video、picture、live 三种类型。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "description": "内容 id（视频号 object id）" },
                  "contact": {
                    "type": "object",
                    "properties": {
                      "username": { "type": "string", "description": "视频号账号 username" },
                      "nickname": { "type": "string", "description": "视频号昵称" },
                      "headUrl": { "type": "string", "description": "视频号头像 URL" }
                    }
                  },
                  "objectDesc": {
                    "type": "object",
                    "properties": {
                      "description": { "type": "string", "description": "标题/描述" },
                      "mediaType": { "type": "integer", "description": "媒体类型（2=图片视频, 4=普通视频, 9=直播）" },
                      "media": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "url": { "type": "string", "description": "媒体 URL" },
                            "urlToken": { "type": "string", "description": "媒体 token" },
                            "coverUrl": { "type": "string", "description": "封面 URL" },
                            "decodeKey": { "type": "string", "description": "解密 key" },
                            "spec": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "fileFormat": { "type": "string", "description": "视频规格名称" }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  },
                  "objectNonceId": { "type": "string", "description": "nonce id" },
                  "source_url": { "type": "string", "description": "来源页面 URL（提供则直接使用）" },
                  "type": { "type": "string", "description": "feed 类型：media、picture、live" },
                  "liveInfo": { "type": "object", "description": "直播信息（live 类型时使用）" },
                  "files": { "type": "array", "description": "图片列表（picture 类型时使用）" },
                  "anchorContact": { "type": "object", "description": "主播信息（live 类型时使用）" }
                },
                "required": ["id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建结果",
            "content": {
              "application/json": {
                "examples": {
                  "success": { "summary": "成功", "value": { "code": 0, "msg": "成功", "data": { "message": "创建下载任务成功", "task_id": "fqG2OhiJFRYpnELqZmPRW" } } },
                  "invalid": { "summary": "参数不合法", "value": { "code": 400, "msg": "不合法的参数", "data": {} } }
                }
              }
            }
          }
        }
      }
    },
    "/api/download_task/batch_create": {
      "post": {
        "tags": ["下载任务"],
        "summary": "批量创建下载任务（兼容 wx_channels_api）",
        "description": "批量接收 ChannelsObject 结构的原始视频号数据数组，后端自动处理 format_feed、链接拼接、文件名生成等。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "description": "内容 id" },
                    "contact": {
                      "type": "object",
                      "properties": {
                        "username": { "type": "string", "description": "视频号账号 username" },
                        "nickname": { "type": "string", "description": "视频号昵称" },
                        "headUrl": { "type": "string", "description": "视频号头像 URL" }
                      }
                    },
                    "objectDesc": {
                      "type": "object",
                      "properties": {
                        "description": { "type": "string", "description": "标题/描述" },
                        "mediaType": { "type": "integer", "description": "媒体类型" },
                        "media": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "url": { "type": "string", "description": "媒体 URL" },
                              "urlToken": { "type": "string", "description": "媒体 token" },
                              "coverUrl": { "type": "string", "description": "封面 URL" },
                              "decodeKey": { "type": "string", "description": "解密 key" }
                            }
                          }
                        }
                      }
                    },
                    "objectNonceId": { "type": "string", "description": "nonce id" },
                    "source_url": { "type": "string", "description": "来源页面 URL" }
                  },
                  "required": ["id"]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建结果",
            "content": {
              "application/json": {
                "examples": {
                  "success": { "summary": "成功", "value": { "code": 0, "msg": "成功", "data": { "ids": ["fqG2OhiJFRYpnELqZmPRW"] } } }
                }
              }
            }
          }
        }
      }
    },
    "/api/download_task/delete": {
      "post": {
        "tags": ["下载任务"],
        "summary": "删除下载任务（兼容 wx_channels_api）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "download_task_id": { "type": "integer", "description": "download_task 表主键 id" } },
                "required": ["download_task_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除结果",
            "content": {
              "application/json": {
                "examples": {
                  "success": { "summary": "成功", "value": { "code": 0, "msg": "成功", "data": { "message": "删除下载任务成功" } } },
                  "not_found": { "summary": "任务不存在", "value": { "code": 404, "msg": "未找到下载任务", "data": {} } }
                }
              }
            }
          }
        }
      }
    },
    "/api/download_task/retry": {
      "post": {
        "tags": ["下载任务"],
        "summary": "重试下载任务（兼容 wx_channels_api）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "id": { "type": "integer" } }, "required": ["id"] }
            }
          }
        },
        "responses": {
          "200": {
            "description": "重试结果",
            "content": {
              "application/json": {
                "examples": {
                  "success": { "summary": "成功", "value": { "code": 0, "msg": "成功", "data": { "message": "已重试下载", "new_task_id": "new_task_id_xxx" } } }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/download_task/highlight_file": {
      "post": {
        "tags": ["下载任务"],
        "summary": "在文件管理器中高亮已下载文件（兼容 wx_channels_api）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "file_path": { "type": "string", "description": "相对下载目录的文件路径" } }, "required": ["file_path"] }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作结果",
            "content": {
              "application/json": {
                "examples": {
                  "success": { "summary": "成功", "value": { "code": 0, "msg": "成功", "data": "Success" } }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/channels/task/status": {
      "get": {
        "tags": ["平台"],
        "summary": "查询任务状态（兼容 wx_channels_api）",
        "parameters": [{ "name": "task_id", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "状态信息" } }
      }
    },
...
  }
}